run_tests.py: various fixes improving robustness (#986)
* Works correctly with test-run timeout and failure.
* Launch correctly shutdowns on SIGINT.
* Move launcher/__init__.py into launch.
* run_tests.py runs with unbuffered output.
