+++ .gdbinit-test
@@ -3,21 +3,26 @@ set output-radix 0x10
 set pagination off
 set confirm off
 set verbose off
+handle SIGINT stop
 
 # make extra commands available
 python import os, sys
 python sys.path.append(os.path.join(os.getcwd(), 'sys'))
 python import debug
 
+define post-mortem
+  info registers
+  # cpu tlb # XXX it sometimes hangs
+  backtrace full
+  kthread *
+  kproc
+  klog
+end
+
 # launch kernel tests and wait for ktest_main to finish
 break panic_fail
 commands
-info registers
-# cpu tlb # XXX it sometimes hangs
-backtrace full
-kthread *
-kproc
-klog
+post-mortem
 kill
 quit 1
 end
@@ -28,4 +33,4 @@ kill
 quit 0
 end
 
-cont
+continue
